--- a/hw10_program_optimization/stats_optimization.txt
+++ b/hw10_program_optimization/stats_optimization.txt
@@ -1,0 +1,16 @@
+goos: linux
+goarch: amd64
+pkg: github.com/qeeq72/hw-test/hw10_program_optimization
+cpu: Intel(R) Core(TM) i7-10875H CPU @ 2.30GHz
+BenchmarkGetDomainStat-4   	     134	   7619291 ns/op	  200660 B/op	    6991 allocs/op
+BenchmarkGetDomainStat-4   	     144	   7203243 ns/op	  187104 B/op	    6506 allocs/op
+BenchmarkGetDomainStat-4   	     146	   7201062 ns/op	  184613 B/op	    6418 allocs/op
+BenchmarkGetDomainStat-4   	     148	   7195038 ns/op	  182189 B/op	    6331 allocs/op
+BenchmarkGetDomainStat-4   	     150	   7692468 ns/op	  179841 B/op	    6247 allocs/op
+BenchmarkGetDomainStat-4   	     139	   7947332 ns/op	  193635 B/op	    6740 allocs/op
+BenchmarkGetDomainStat-4   	     132	   8192745 ns/op	  203630 B/op	    7096 allocs/op
+BenchmarkGetDomainStat-4   	     147	   7257608 ns/op	  183392 B/op	    6374 allocs/op
+BenchmarkGetDomainStat-4   	     140	   7260005 ns/op	  192294 B/op	    6692 allocs/op
+BenchmarkGetDomainStat-4   	     142	   7260273 ns/op	  189660 B/op	    6598 allocs/op
+PASS
+ok  	github.com/qeeq72/hw-test/hw10_program_optimization	33.165s