--- a/hw10_program_optimization/stats_optimization.txt
+++ b/hw10_program_optimization/stats_optimization.txt
@@ -2,15 +2,15 @@ goos: linux
 goarch: amd64
 pkg: github.com/qeeq72/hw-test/hw10_program_optimization
 cpu: Intel(R) Core(TM) i7-10875H CPU @ 2.30GHz
-BenchmarkGetDomainStat-4   	     134	   7619291 ns/op	  200660 B/op	    6991 allocs/op
-BenchmarkGetDomainStat-4   	     144	   7203243 ns/op	  187104 B/op	    6506 allocs/op
-BenchmarkGetDomainStat-4   	     146	   7201062 ns/op	  184613 B/op	    6418 allocs/op
-BenchmarkGetDomainStat-4   	     148	   7195038 ns/op	  182189 B/op	    6331 allocs/op
-BenchmarkGetDomainStat-4   	     150	   7692468 ns/op	  179841 B/op	    6247 allocs/op
-BenchmarkGetDomainStat-4   	     139	   7947332 ns/op	  193635 B/op	    6740 allocs/op
-BenchmarkGetDomainStat-4   	     132	   8192745 ns/op	  203630 B/op	    7096 allocs/op
-BenchmarkGetDomainStat-4   	     147	   7257608 ns/op	  183392 B/op	    6374 allocs/op
-BenchmarkGetDomainStat-4   	     140	   7260005 ns/op	  192294 B/op	    6692 allocs/op
-BenchmarkGetDomainStat-4   	     142	   7260273 ns/op	  189660 B/op	    6598 allocs/op
+BenchmarkGetDomainStat-4   	     140	   7250015 ns/op	  162470 B/op	    6702 allocs/op
+BenchmarkGetDomainStat-4   	     147	   7324018 ns/op	  155289 B/op	    6385 allocs/op
+BenchmarkGetDomainStat-4   	     160	   7102042 ns/op	  143176 B/op	    5868 allocs/op
+BenchmarkGetDomainStat-4   	     144	   7288698 ns/op	  158648 B/op	    6517 allocs/op
+BenchmarkGetDomainStat-4   	     144	   7332494 ns/op	  158128 B/op	    6517 allocs/op
+BenchmarkGetDomainStat-4   	     147	   7238291 ns/op	  155042 B/op	    6385 allocs/op
+BenchmarkGetDomainStat-4   	     147	   7176539 ns/op	  155038 B/op	    6385 allocs/op
+BenchmarkGetDomainStat-4   	     146	   7225219 ns/op	  156046 B/op	    6428 allocs/op
+BenchmarkGetDomainStat-4   	     150	   7151628 ns/op	  152311 B/op	    6258 allocs/op
+BenchmarkGetDomainStat-4   	     148	   8497761 ns/op	  154023 B/op	    6342 allocs/op
 PASS
-ok  	github.com/qeeq72/hw-test/hw10_program_optimization	33.165s
+ok  	github.com/qeeq72/hw-test/hw10_program_optimization	34.446s