@qeeq72 2021-12-16
Новые результаты benchstat:

name             old time/op    new time/op    delta
GetDomainStat-4     629ms ±87%       7ms ± 2%  -98.85%  (p=0.000 n=10+9)

name             old alloc/op   new alloc/op   delta
GetDomainStat-4     179MB ±71%       0MB ± 4%  -99.91%  (p=0.000 n=10+9)

name             old allocs/op  new allocs/op  delta
GetDomainStat-4     1.83M ±71%     0.01M ± 4%  -99.65%  (p=0.000 n=10+9)